pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-cost_aware_finqa"
version = "0.1.0"
description = "Cost-Aware FinQA: An RL environment for training agents to make strategic tool-use decisions in financial research"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "gradio>=4.0.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "cost_aware_finqa.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["cost_aware_finqa", "cost_aware_finqa.server"]
package-dir = { "cost_aware_finqa" = ".", "cost_aware_finqa.server" = "server" }

[tool.setuptools.package-data]
"cost_aware_finqa" = ["data/*.json", "data/*.db"]
